requirements.txt
websocket-client
gspread
oauth2client
gspread
ccxt